fix:[UIUX-600] fixed compliance % showing 100-100 always
