+++ webapp/src/app/shared/services/filter-management.service.ts
@@ -420,6 +420,28 @@ import { WorkflowService } from 'src/app/core/services/workflow.service';
        shouldUpdateFilters, shouldUpdateData, filterText, preApply
      };
    }
+
+   getRangeFilterOptions (filterTagsData, addLastOptionSeparate = false) {
+     const numOfIntervals = 5;
+     const { min, max } = filterTagsData.optionRange;
+     const intervals = this.utils.generateIntervals(min, max, numOfIntervals);
+
+     filterTagsData = [];
+     let includeLastOption = false;
+     intervals.forEach(interval => {
+       const lb = Math.round(interval.lowerBound);
+       let up = Math.round(interval.upperBound);
+       includeLastOption = up === 100;
+       if (addLastOptionSeparate && up === 100 && lb !== up) {
+         up--;
+       }
+       filterTagsData.push({ id: `${lb}-${up}`, name: `${lb}-${up}` });
+     });
+     if (addLastOptionSeparate && includeLastOption) {
+       filterTagsData.push({ id: "100-100", name: "100-100" });
+     }
+     return filterTagsData;
+   }
  
  }
